+++ src/core/FishPeerSend.cpp
@@ -7,7 +7,7 @@
  * Stores reliable datagrams for retransmission until ACK'd.
  */
 
-#include "fishnet/FishPeer.h"
+#include <fishnet/FishPeer.h>
 #include <algorithm>
 #include <cstdio>